refector: TeamArray 관련 로직을 옮겼습니다
